fix(web): migrate layout in to root file
