+++ apps/web/app/root.tsx
@@ -1,9 +1,20 @@
+import { Links, Meta, Scripts, ScrollRestoration } from 'react-router';
+
+import { css } from 'leather-styles/css';
+import { Box, Flex, styled } from 'leather-styles/jsx';
 import { Buffer } from 'safe-buffer';
 
+import { HasChildren } from '@leather.io/ui';
 import leatherUiStyles from '@leather.io/ui/styles?url';
 
 import type { Route } from './+types/root';
 import stylesheet from './app.css?url';
+import { defaultMetaTags } from './constants/meta-tags';
+import { InstallDialog } from './features/install-dialog/install-dialog';
+import { MockLeatherDialog } from './features/mock-dialog/mock-dialog';
+import { Footer } from './layouts/footer/footer';
+import { GlobalLoader } from './layouts/nav/global-loader';
+import { Nav } from './layouts/nav/nav';
 import { ErrorPage } from './layouts/page/error';
 
 // Polyfill global Buffer
@@ -20,4 +31,45 @@ export function ErrorBoundary({ error }: Route.ErrorBoundaryProps) {
   return <ErrorPage error={error} />;
 }
 
-export { default, Layout } from './app';
+const maxWidthCss = css({
+  maxWidth: '1200px',
+  margin: '0 auto',
+  width: '100%',
+});
+
+export function Layout({ children }: HasChildren) {
+  return (
+    <html lang="en">
+      <head>
+        <Meta />
+        {defaultMetaTags.map((meta, i) => (
+          <meta key={'meta' + i} {...meta} />
+        ))}
+        <Links />
+      </head>
+      <styled.body>
+        <GlobalLoader />
+        <Nav />
+        <Flex
+          flexDir="column"
+          marginLeft={[null, null, 'navbar']}
+          minHeight="100vh"
+          px={['space.04', null, 'space.07']}
+        >
+          <styled.main flex={1} bg="ink.background-primary" className={maxWidthCss}>
+            {children}
+          </styled.main>
+          <Box className={maxWidthCss}>
+            <Footer />
+          </Box>
+        </Flex>
+        <InstallDialog />
+        <MockLeatherDialog />
+        <ScrollRestoration />
+        <Scripts />
+      </styled.body>
+    </html>
+  );
+}
+
+export { default } from './app';
