+++ apps/web/package.json
@@ -7,7 +7,7 @@
     "build:watch": "pnpm dev",
     "cf-typegen": "wrangler types",
     "deploy": "pnpm build && wrangler deploy",
-    "dev": "NODE_OPTIONS='--import ./instrument.server.mjs' react-router dev",
+    "dev": "react-router dev",
     "format": "prettier --write \"app/**/*.{ts,tsx}\" --ignore-path ../../.prettierignore",
     "format:check": "prettier --check \"app/**/*.{ts,tsx}\" --ignore-path ../../.prettierignore",
     "lint": "eslint \"app/**/*.{ts,tsx}\"",
@@ -84,14 +84,14 @@
     "zod": "4.0.17"
   },
   "devDependencies": {
-    "@cloudflare/vite-plugin": "1.13.12",
-    "@cloudflare/workers-types": "4.20251011.0",
+    "@cloudflare/vite-plugin": "1.15.2",
+    "@cloudflare/workers-types": "4.20251121.0",
     "@msw/playwright": "0.4.2",
     "@originjs/vite-plugin-commonjs": "1.0.3",
     "@pandacss/dev": "0.53.6",
     "@playwright/test": "1.56.0",
-    "@react-router/cloudflare": "7.9.4",
-    "@react-router/dev": "7.9.4",
+    "@react-router/cloudflare": "7.9.6",
+    "@react-router/dev": "7.9.6",
     "@types/d3": "7.4.3",
     "@types/node": "24.0.8",
     "@types/nprogress": "0.2.3",
@@ -101,7 +101,7 @@
     "@types/valid-url": "1.0.7",
     "jsdom": "26.1.0",
     "msw": "2.10.4",
-    "react-router-devtools": "5.1.3",
+    "react-router-devtools": "6.0.0",
     "rollup-plugin-node-polyfills": "0.2.1",
     "ts-node": "10.9.2",
     "tty-browserify": "0.0.1",
@@ -110,6 +110,6 @@
     "vite-plugin-svgr": "4.3.0",
     "vite-tsconfig-paths": "5.1.4",
     "vitest": "2.1.9",
-    "wrangler": "4.42.2"
+    "wrangler": "4.50.0"
   }
 }
